changed map to comply with condition 2/3 of whole map is ocean
apparently it's actually 2/3 OF THE WHOLE MAP and not the ratio of oceans to non oceans >2/3 therefore we could have multiple modes to fill the map.
We thought of 4 different modes and coded the first easiest one.
The 4th mode have a lot of flexibility and is the last goal if there is enough time.
We decided to the pass the mode choice as an argument to the Map public constructor, instead of making it a global variable inside the class map.