fixed bug unit[i] != null in player's units array after killing it.
we can't pass an object by reference. so we set the owner of the unit to null and loop through all the units from the class player and whenever the owner == null we set the unit[i] object to null.
still need to fix bug 3 soldiers on sand tile only take 5 food units and not 6