added exchange system for WarPlayer
exchange system is working but need to fix bugs in the order of actions: defining how many food units will be eaten by this unit,  defining how many resources will be produced by this unit, putting the unit on the tile, harvesting, sending the resources to the player and the end of the turn. in the next turn start off by eating then repeat until unit death or end of game.